changed the way files are scanned to have a more fine control added issues concept and reflect that in the UI changed the logic of new versions and bump version removed spinner and replace with progress bar for console fixed many issues (and probably adding new ones)

+++ process/organizefolderStructure.go
@@ -16,14 +16,15 @@ import (
 )
 
 var (
-	folderIllegalCharsRegex = regexp.MustCompile(`[/\\?%*:|"<>]`)
+	folderIllegalCharsRegex = regexp.MustCompile(`[./\\?%*:;=|"<>]`)
+	nonAscii                = regexp.MustCompile("[[:^ascii:]]")
 )
 
 func DeleteOldUpdates(localDB *db.LocalSwitchFilesDB, updateProgress db.ProgressUpdater) {
 	i := 0
 	for _, v := range localDB.TitlesMap {
 		if updateProgress != nil {
-			updateProgress.UpdateProgress(i, len(localDB.TitlesMap), v.File.Info.Name())
+			updateProgress.UpdateProgress(i, len(localDB.TitlesMap), v.File.ExtendedInfo.Info.Name())
 		}
 		if len(v.Updates) > 1 {
 			//sort the available local versions
@@ -40,14 +41,14 @@ func DeleteOldUpdates(localDB *db.LocalSwitchFilesDB, updateProgress db.Progress
 					//should not happen, but make sure we do not delete base
 					continue
 				}
-				fileToRemove := filepath.Join(v.Updates[localVersions[i]].BaseFolder, v.Updates[localVersions[i]].Info.Name())
+				fileToRemove := filepath.Join(v.Updates[localVersions[i]].ExtendedInfo.BaseFolder, v.Updates[localVersions[i]].ExtendedInfo.Info.Name())
 				zap.S().Infof("--> [Delete] Old update file: %v [latest update:%v]\n", fileToRemove, localVersions[len(localVersions)-1])
 				err := os.Remove(fileToRemove)
 				if err != nil {
 					zap.S().Errorf("Failed to delete file  %v  [%v]\n", fileToRemove, err)
 				}
 			}
-			v.Updates = map[int]db.ExtendedFileInfo{localVersions[len(localVersions)-1]: v.Updates[localVersions[len(localVersions)-1]]}
+			v.Updates = map[int]db.SwitchFileInfo{localVersions[len(localVersions)-1]: v.Updates[localVersions[len(localVersions)-1]]}
 		}
 
 	}
@@ -73,7 +74,7 @@ func OrganizeByFolders(baseFolder string,
 			continue
 		}
 		if updateProgress != nil {
-			updateProgress.UpdateProgress(i, tasksSize, v.File.Info.Name())
+			updateProgress.UpdateProgress(i, tasksSize, v.File.ExtendedInfo.Info.Name())
 		}
 
 		titleName := getTitleName(titlesDB.TitlesMap[k], v)
@@ -85,7 +86,7 @@ func OrganizeByFolders(baseFolder string,
 		templateData[settings.TEMPLATE_TITLE_NAME] = titleName
 		templateData[settings.TEMPLATE_VERSION] = "0"
 
-		var destinationPath = v.File.BaseFolder
+		var destinationPath = v.File.ExtendedInfo.BaseFolder
 
 		//create folder if needed
 		if options.CreateFolderPerGame {
@@ -101,8 +102,8 @@ func OrganizeByFolders(baseFolder string,
 		}
 
 		//process base title
-		from := filepath.Join(v.File.BaseFolder, v.File.Info.Name())
-		to := filepath.Join(destinationPath, getFileName(options, v.File.Info.Name(), templateData))
+		from := filepath.Join(v.File.ExtendedInfo.BaseFolder, v.File.ExtendedInfo.Info.Name())
+		to := filepath.Join(destinationPath, getFileName(options, v.File.ExtendedInfo.Info.Name(), templateData))
 		err := moveFile(from, to)
 		if err != nil {
 			zap.S().Errorf("Failed to move file [%v]\n", err)
@@ -116,11 +117,11 @@ func OrganizeByFolders(baseFolder string,
 			}
 			templateData[settings.TEMPLATE_VERSION] = strconv.Itoa(update)
 			templateData[settings.TEMPLATE_TYPE] = "UPD"
-			from = filepath.Join(updateInfo.BaseFolder, updateInfo.Info.Name())
+			from = filepath.Join(updateInfo.ExtendedInfo.BaseFolder, updateInfo.ExtendedInfo.Info.Name())
 			if options.CreateFolderPerGame {
-				to = filepath.Join(destinationPath, getFileName(options, updateInfo.Info.Name(), templateData))
+				to = filepath.Join(destinationPath, getFileName(options, updateInfo.ExtendedInfo.Info.Name(), templateData))
 			} else {
-				to = filepath.Join(updateInfo.BaseFolder, getFileName(options, updateInfo.Info.Name(), templateData))
+				to = filepath.Join(updateInfo.ExtendedInfo.BaseFolder, getFileName(options, updateInfo.ExtendedInfo.Info.Name(), templateData))
 			}
 			err := moveFile(from, to)
 			if err != nil {
@@ -137,11 +138,11 @@ func OrganizeByFolders(baseFolder string,
 			templateData[settings.TEMPLATE_TYPE] = "DLC"
 			templateData[settings.TEMPLATE_TITLE_ID] = id
 			templateData[settings.TEMPLATE_DLC_NAME] = getDlcName(titlesDB.TitlesMap[k], dlc)
-			from = filepath.Join(dlc.BaseFolder, dlc.Info.Name())
+			from = filepath.Join(dlc.ExtendedInfo.BaseFolder, dlc.ExtendedInfo.Info.Name())
 			if options.CreateFolderPerGame {
-				to = filepath.Join(destinationPath, getFileName(options, dlc.Info.Name(), templateData))
+				to = filepath.Join(destinationPath, getFileName(options, dlc.ExtendedInfo.Info.Name(), templateData))
 			} else {
-				to = filepath.Join(dlc.BaseFolder, getFileName(options, dlc.Info.Name(), templateData))
+				to = filepath.Join(dlc.ExtendedInfo.BaseFolder, getFileName(options, dlc.ExtendedInfo.Info.Name(), templateData))
 			}
 			err = moveFile(from, to)
 			if err != nil {
@@ -200,31 +201,30 @@ func isOptionsValid(options settings.OrganizeOptions) bool {
 	return true
 }
 
-func getDlcName(switchTitle *db.SwitchTitle, file db.ExtendedFileInfo) string {
+func getDlcName(switchTitle *db.SwitchTitle, file db.SwitchFileInfo) string {
 	if switchTitle == nil {
 		return ""
 	}
 	if dlcAttributes, ok := switchTitle.Dlc[file.Metadata.TitleId]; ok {
 		name := dlcAttributes.Name
 		name = strings.ReplaceAll(name, "\n", "")
-		return folderIllegalCharsRegex.ReplaceAllString(name, "-")
+		return name
 	}
 	return ""
 }
 
-func getTitleName(switchTitle *db.SwitchTitle, v *db.SwitchFile) string {
+func getTitleName(switchTitle *db.SwitchTitle, v *db.SwitchGameFiles) string {
 	if switchTitle != nil && switchTitle.Attributes.Name != "" {
-		name := switchTitle.Attributes.Name
-		return folderIllegalCharsRegex.ReplaceAllString(name, "-")
+		return switchTitle.Attributes.Name
 	} else {
 		//for non eshop games (cartridge only), grab the name from the file
-		return db.ParseTitleNameFromFileName(v.File.Info.Name())
+		return db.ParseTitleNameFromFileName(v.File.ExtendedInfo.Info.Name())
 	}
 }
 
 func getFolderName(options settings.OrganizeOptions, templateData map[string]string) string {
 
-	return applyTemplate(templateData, options.FolderNameTemplate)
+	return applyTemplate(templateData, options.SwitchSafeFileNames, options.FolderNameTemplate)
 }
 
 func getFileName(options settings.OrganizeOptions, originalName string, templateData map[string]string) string {
@@ -232,7 +232,7 @@ func getFileName(options settings.OrganizeOptions, originalName string, template
 		return originalName
 	}
 	ext := path.Ext(originalName)
-	result := applyTemplate(templateData, options.FileNameTemplate)
+	result := applyTemplate(templateData, options.SwitchSafeFileNames, options.FileNameTemplate)
 	return result + ext
 }
 
@@ -244,7 +244,7 @@ func moveFile(from string, to string) error {
 	return err
 }
 
-func applyTemplate(templateData map[string]string, template string) string {
+func applyTemplate(templateData map[string]string, useSafeNames bool, template string) string {
 	result := strings.Replace(template, "{"+settings.TEMPLATE_TITLE_NAME+"}", templateData[settings.TEMPLATE_TITLE_NAME], 1)
 	result = strings.Replace(result, "{"+settings.TEMPLATE_TITLE_ID+"}", strings.ToUpper(templateData[settings.TEMPLATE_TITLE_ID]), 1)
 	result = strings.Replace(result, "{"+settings.TEMPLATE_VERSION+"}", templateData[settings.TEMPLATE_VERSION], 1)
@@ -258,21 +258,29 @@ func applyTemplate(templateData map[string]string, template string) string {
 	result = strings.ReplaceAll(result, "[]", "")
 	result = strings.ReplaceAll(result, "()", "")
 	result = strings.ReplaceAll(result, "<>", "")
-	result = strings.ReplaceAll(result, "  ", " ")
-	result = strings.TrimSpace(result)
 	if strings.HasSuffix(result, ".") {
 		result = result[:len(result)-1]
 	}
-	return nihongo.RomajiString(result)
+
+	if useSafeNames {
+		result = nihongo.RomajiString(result)
+		result = nonAscii.ReplaceAllLiteralString(result, "")
+	}
+	result = strings.ReplaceAll(result, "  ", " ")
+	result = strings.TrimSpace(result)
+	return folderIllegalCharsRegex.ReplaceAllString(result, "")
 }
 
 func deleteEmptyFolders(path string) error {
 	err := filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
-			return err
+			zap.S().Error("Error while deleting empty folders", err)
 		}
-		if info.IsDir() {
-			deleteEmptyFolder(path)
+		if info != nil && info.IsDir() {
+			err = deleteEmptyFolder(path)
+			if err != nil {
+				zap.S().Error("Error while deleting empty folders", err)
+			}
 		}
 
 		return nil
